renamed Concept to PreviewConcept to be more similar to PreviewDraft structure + fix markdown rendering
