+++ src/components/PreviewConcept/PreviewConcept.tsx
@@ -7,65 +7,94 @@
  */
 
 import React, { FC } from 'react';
-import { injectT, tType } from '@ndla/i18n';
 import styled from '@emotion/styled';
-import Concept from './Concept';
-import { StyledPreviewHeader } from '../PreviewDraft/PreviewLanguage';
-import StyledPreviewTwoArticles from '../PreviewDraft/StyledPreviewTwoArticles';
-import { Concept as ConceptType } from '../SlateEditor/editorTypes';
+import { spacing } from '@ndla/core';
+import { injectT, tType } from '@ndla/i18n';
+import {
+  NotionDialogContent,
+  NotionHeaderWithoutExitButton,
+  NotionDialogLicenses,
+  NotionDialogText,
+} from '@ndla/notion';
+import { Remarkable } from 'remarkable';
+import { Concept } from '../SlateEditor/editorTypes';
+import { getSrcSets } from '../../util/imageEditorUtil';
 
-const StyledPreviewConcept = styled('div')`
-  display: inline-block;
-  width: 100%;
+const StyledBody = styled.div`
+  margin: 0 ${spacing.normal} ${spacing.small};
+  .c-tabs {
+    margin-left: 0;
+    margin-right: 0;
+  }
+  .react-tabs__tab-panel--selected {
+    animation-name: fadeInLeft;
+    animation-duration: 500ms;
+  }
 `;
 
 interface Props {
-  firstConcept: ConceptType;
-  secondConcept: ConceptType;
-  onChangePreviewLanguage: Function;
-  previewLanguage: string;
+  concept: Concept;
 }
 
-const PreviewConcept: FC<Props & tType> = ({
-  t,
-  firstConcept,
-  secondConcept,
-  onChangePreviewLanguage,
-  previewLanguage,
-}) => {
+const PreviewConcept: FC<Props & tType> = ({ concept }) => {
+  const markdown = new Remarkable({ breaks: true });
+  markdown.inline.ruler.enable(['sub', 'sup']);
+  markdown.block.ruler.disable(['list']);
+
+  const VisualElement = () => {
+    const visualElement = concept.visualElement;
+    switch (visualElement?.resource) {
+      case 'image':
+        const srcSet = getSrcSets(visualElement.resource_id, visualElement);
+        return (
+          <img
+            alt={visualElement?.alt}
+            src={visualElement?.url}
+            srcSet={srcSet}
+          />
+        );
+      case 'video':
+      case 'brightcove':
+      case 'external':
+      case 'h5p':
+        return (
+          <iframe
+            title={visualElement?.title}
+            src={visualElement?.url}
+            frameBorder="0"
+            scrolling="no"
+            height={400}
+          />
+        );
+      default:
+        return null;
+    }
+  };
+
   return (
-    <StyledPreviewConcept>
-      <StyledPreviewTwoArticles>
-        <StyledPreviewHeader>
-          <h2 className="u-4/6@desktop u-push-1/6@desktop">
-            {t('form.previewLanguageArticle.title', {
-              language: t(`language.${firstConcept.language}`).toLowerCase(),
-            })}
-          </h2>
-        </StyledPreviewHeader>
-        <Concept concept={firstConcept} />
-      </StyledPreviewTwoArticles>
-      <StyledPreviewTwoArticles>
-        <StyledPreviewHeader>
-          <h2 className="u-4/6@desktop u-push-1/6@desktop">
-            {t('form.previewLanguageArticle.title', {
-              language: t(`language.${previewLanguage}`).toLowerCase(),
-            })}
-          </h2>
-          <select
-            className="u-4/6@desktop u-push-1/6@desktop"
-            onChange={evt => onChangePreviewLanguage(evt.target.value)}
-            value={previewLanguage}>
-            {firstConcept.supportedLanguages.map(language => (
-              <option key={language} value={language}>
-                {t(`language.${language}`)}
-              </option>
-            ))}
-          </select>
-        </StyledPreviewHeader>
-        <Concept concept={secondConcept} />
-      </StyledPreviewTwoArticles>
-    </StyledPreviewConcept>
+    <>
+      <NotionHeaderWithoutExitButton
+        title={concept.title}
+        subtitle={concept.metaDescription}
+      />
+      <StyledBody>
+        <NotionDialogContent>
+          <VisualElement />
+          <NotionDialogText>
+            <span
+              dangerouslySetInnerHTML={{
+                __html: markdown.render(concept.content),
+              }}
+            />
+          </NotionDialogText>
+        </NotionDialogContent>
+        <NotionDialogLicenses
+          license={concept.copyright?.license?.license}
+          source={concept.source}
+          authors={concept.copyright?.creators.map(creator => creator.name)}
+        />
+      </StyledBody>
+    </>
   );
 };
